Push two more errors into the archivist
