+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/bag/BagDigestFileCreator.scala
@@ -1,13 +1,11 @@
 package uk.ac.wellcome.platform.archive.archivist.bag
 
+import uk.ac.wellcome.platform.archive.archivist.models.errors.InvalidBagManifestError
 import uk.ac.wellcome.platform.archive.common.models.bagit.{
   BagDigestFile,
   BagItemPath
 }
-import uk.ac.wellcome.platform.archive.common.models.error.{
-  ArchiveError,
-  InvalidBagManifestError
-}
+import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
 
 import scala.util.matching.Regex
 import scala.util.{Failure, Success, Try}
