+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/DownloadAndVerifyDigestItemFlow.scala
@@ -8,11 +8,11 @@ import akka.stream.scaladsl.{Flow, Source, StreamConverters}
 import com.amazonaws.services.s3.AmazonS3
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.platform.archive.archivist.models.DigestItemJob
-import uk.ac.wellcome.platform.archive.archivist.models.errors.ChecksumNotMatchedOnDownloadError
-import uk.ac.wellcome.platform.archive.common.models.error.{
-  ArchiveError,
+import uk.ac.wellcome.platform.archive.archivist.models.errors.{
+  ChecksumNotMatchedOnDownloadError,
   DownloadError
 }
+import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
 import uk.ac.wellcome.storage.ObjectLocation
 
 import scala.util.{Failure, Success, Try}
